PNE-6482 Support `default_labels` in MaterialSpec constructor
This update is analagous to the one in https://github.com/CitrineInformatics/citrine-python/pull/989

The associated utility function and test have been moved to a shared module,
and the documentation has been updated
