+++ src/citrine/__version__.py
@@ -1,1 +1,1 @@
-__version__ = "3.18.0"
+__version__ = "3.19.0"
